Split out isPostsLastPageForQuery
